Share some tests between all versions, add easy constructors for valuesets.
