+++ src/Hl7.Fhir.Specification.Shared.Tests/Hl7.Fhir.Specification.Shared.Tests.projitems
@@ -39,6 +39,8 @@
     <Compile Include="$(MSBuildThisFileDirectory)StructureDefinitionSummaryProviderTest.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)StructureDefinitionWalkerTests.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Terminology\LocalTerminologyServiceTests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Terminology\SharedExpansionTests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Terminology\ValueSetExtensionsTests.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)TestDataHelper.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)TimingSource.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)xUnitBug.cs" />
